solids - reduce to minimal example set, see Ratel for full
