+++ examples/solids/index.md
@@ -737,43 +737,3 @@ $$ (tau-neo-hookean-stable)
 
 which is more numerically stable for small strain, and thus preferred for computation. Note that the $\log J$ is computed via `log1p` {eq}`log1p`, as we discussed in the previous tip.
 :::
-
-### Jacobian representation
-
-We have implemented four storage variants for the Jacobian in our finite strain hyperelasticity. In each case, some variables are computed during residual evaluation and used during Jacobian application.
-
-:::{list-table} Four algorithms for Jacobian action in finite strain hyperelasticity problem
-:header-rows: 1
-:widths: auto
-
-* - Option `-problem`
-  - Static storage
-  - Computed storage
-  - \# scalars
-  - Equations
-
-
-* - `FSInitial-NH1`
-  - $\nabla_{X} \hat X, \operatorname{det}\nabla_{\hat X} X$
-  - $\nabla_X \bm u$
-  - 19
-  - {eq}`eq-diff-P` {eq}`eq-neo-hookean-incremental-stress`
-
-* - `FSInitial-NH2`
-  - $\nabla_{X} \hat X, \operatorname{det}\nabla_{\hat X} X$
-  - $\nabla_X \bm u, \bm C^{-1}, \lambda \log J$
-  - 26
-  - {eq}`eq-diff-P` {eq}`eq-neo-hookean-incremental-stress`
-
-* - `FSCurrent-NH1`
-  - $\nabla_{X} \hat X, \operatorname{det}\nabla_{\hat X} X$
-  - $\nabla_X \bm u$
-  - 19
-  - {eq}`jacobian-weak-form-current` {eq}`nabla_xdu`
-
-* - `FSCurrent-NH2`
-  - $\operatorname{det}\nabla_{\hat X} X$
-  - $\nabla_x \hat X, \bm \tau, \lambda \log J$
-  - 17
-  - {eq}`jacobian-weak-form-current` {eq}`jacobian-weak-form-current2`
-:::
