+++ examples/solids/problems/linear.c
@@ -22,15 +22,15 @@ ProblemData linear_elasticity = {
     .setup_geo_loc        = SetupGeo_loc,
     .q_data_size          = 10,
     .quadrature_mode      = CEED_GAUSS,
-    .residual             = ElasLinearF,
-    .residual_loc         = ElasLinearF_loc,
+    .residual             = ElasResidual_Linear,
+    .residual_loc         = ElasResidual_Linear_loc,
     .number_fields_stored = 0,
-    .jacobian             = ElasLineardF,
-    .jacobian_loc         = ElasLineardF_loc,
-    .energy               = ElasLinearEnergy,
-    .energy_loc           = ElasLinearEnergy_loc,
-    .diagnostic           = ElasLinearDiagnostic,
-    .diagnostic_loc       = ElasLinearDiagnostic_loc,
+    .jacobian             = ElasJacobian_Linear,
+    .jacobian_loc         = ElasJacobian_Linear_loc,
+    .energy               = ElasEnergy_Linear,
+    .energy_loc           = ElasEnergy_Linear_loc,
+    .diagnostic           = ElasDiagnostic_Linear,
+    .diagnostic_loc       = ElasDiagnostic_Linear_loc,
     .true_soln            = MMSTrueSoln,
     .true_soln_loc        = MMSTrueSoln_loc,
 };
