+++ examples/solids/problems/problems.c
@@ -15,11 +15,8 @@ PetscErrorCode RegisterProblems(ProblemFunctions problem_functions) {
 
   SOLIDS_PROBLEM_REGISTER(problem_functions, "Linear", ElasLinear, NH);
   SOLIDS_PROBLEM_REGISTER(problem_functions, "SS-NH", ElasSSNH, NH);
-  SOLIDS_PROBLEM_REGISTER(problem_functions, "FSCurrent-NH1", ElasFSCurrentNH1, NH);
-  SOLIDS_PROBLEM_REGISTER(problem_functions, "FSCurrent-NH2", ElasFSCurrentNH2, NH);
-  SOLIDS_PROBLEM_REGISTER(problem_functions, "FSInitial-NH1", ElasFSInitialNH1, NH);
-  SOLIDS_PROBLEM_REGISTER(problem_functions, "FSInitial-NH2", ElasFSInitialNH2, NH);
-  SOLIDS_PROBLEM_REGISTER(problem_functions, "FSInitial-MR1", ElasFSInitialMR1, MR);
+  SOLIDS_PROBLEM_REGISTER(problem_functions, "FS-NH", ElasFSNH, NH);
+  SOLIDS_PROBLEM_REGISTER(problem_functions, "FS-MR", ElasFSMR, MR);
 
   PetscFunctionReturn(PETSC_SUCCESS);
 };
