+++ examples/solids/qfunctions/finite-strain-mooney-rivlin.h
@@ -11,12 +11,17 @@
 #include <ceed.h>
 #include <math.h>
 
-#ifndef PHYSICS_STRUCT
-#define PHYSICS_STRUCT
-typedef struct Physics_private *Physics;
-struct Physics_private {
-  CeedScalar nu;  // Poisson's ratio
-  CeedScalar E;   // Young's Modulus
+// -----------------------------------------------------------------------------
+// Mooney-Rivlin context
+#ifndef PHYSICS_STRUCT_MR
+#define PHYSICS_STRUCT_MR
+typedef struct Physics_private_MR *Physics_MR;
+
+struct Physics_private_MR {
+  // material properties for MR
+  CeedScalar mu_1;
+  CeedScalar mu_2;
+  CeedScalar lambda;
 };
 #endif
 
@@ -86,15 +91,14 @@ CEED_QFUNCTION_HELPER int computeMatinvSym(const CeedScalar A[3][3], const CeedS
   };
   for (CeedInt m = 0; m < 6; m++) Ainv[m] = B[m] / (detA);
 
-  return 0;
-};
+  return CEED_ERROR_SUCCESS;
+}
 #endif
-
 // -----------------------------------------------------------------------------
 // Common computations between FS and dFS
 // -----------------------------------------------------------------------------
-CEED_QFUNCTION_HELPER int commonFS(const CeedScalar lambda, const CeedScalar mu, const CeedScalar grad_u[3][3], CeedScalar Swork[6],
-                                   CeedScalar Cinvwork[6], CeedScalar *logJ) {
+CEED_QFUNCTION_HELPER int commonFSMR(const CeedScalar mu_1, const CeedScalar mu_2, const CeedScalar lambda, const CeedScalar grad_u[3][3],
+                                     CeedScalar Swork[6], CeedScalar Cwork[6], CeedScalar Cinvwork[6], CeedScalar *logJ) {
   // E - Green-Lagrange strain tensor
   //     E = 1/2 (grad_u + grad_u^T + grad_u^T*grad_u)
   const CeedInt indj[6] = {0, 1, 2, 1, 0, 0}, indk[6] = {0, 1, 2, 2, 2, 1};
@@ -108,8 +112,6 @@ CEED_QFUNCTION_HELPER int commonFS(const CeedScalar lambda, const CeedScalar mu,
       {E2work[5], E2work[1], E2work[3]},
       {E2work[4], E2work[3], E2work[2]}
   };
-  // J-1
-  const CeedScalar Jm1 = computeJM1(grad_u);
 
   // C : right Cauchy-Green tensor
   // C = I + 2E
@@ -119,30 +121,39 @@ CEED_QFUNCTION_HELPER int commonFS(const CeedScalar lambda, const CeedScalar mu,
       {E2[0][2],     E2[1][2],     1 + E2[2][2]}
   };
 
+  Cwork[0] = C[0][0];
+  Cwork[1] = C[1][1];
+  Cwork[2] = C[2][2];
+  Cwork[3] = C[1][2];
+  Cwork[4] = C[0][2];
+  Cwork[5] = C[0][1];
+  // Compute invariants
+  // I_1 = trace(C)
+  const CeedScalar I_1 = C[0][0] + C[1][1] + C[2][2];
+  // J-1
+  const CeedScalar Jm1 = computeJM1(grad_u);
+  // J = Jm1 + 1
   // Compute C^(-1) : C-Inverse
   const CeedScalar detC = (Jm1 + 1.) * (Jm1 + 1.);
   computeMatinvSym(C, detC, Cinvwork);
 
-  const CeedScalar C_inv[3][3] = {
-      {Cinvwork[0], Cinvwork[5], Cinvwork[4]},
-      {Cinvwork[5], Cinvwork[1], Cinvwork[3]},
-      {Cinvwork[4], Cinvwork[3], Cinvwork[2]}
-  };
-
   // Compute the Second Piola-Kirchhoff (S)
+  // S = (lambda*logJ - mu_1 -2*mu_2)*Cinvwork +(mu_1+mu_2*I_1)*I3-mu_2*Cwork
+  // *1 for indices 0-2 for I_3
+
   *logJ = log1p_series_shifted(Jm1);
-  for (CeedInt m = 0; m < 6; m++) {
-    Swork[m] = (lambda * (*logJ)) * Cinvwork[m];
-    for (CeedInt n = 0; n < 3; n++) Swork[m] += mu * C_inv[indj[m]][n] * E2[n][indk[m]];
+  for (CeedInt i = 0; i < 6; i++) {
+    Swork[i] = (lambda * *logJ - mu_1 - 2 * mu_2) * Cinvwork[i] + (mu_1 + mu_2 * I_1) * (i < 3)  // identity I_3
+               - mu_2 * Cwork[i];
   }
 
-  return 0;
-};
+  return CEED_ERROR_SUCCESS;
+}
 
 // -----------------------------------------------------------------------------
 // Residual evaluation for hyperelasticity, finite strain
 // -----------------------------------------------------------------------------
-CEED_QFUNCTION(ElasFSInitialNH1F)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
+CEED_QFUNCTION(ElasFSResidual_MR)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
   // Inputs
   const CeedScalar(*ug)[3][CEED_Q_VLA] = (const CeedScalar(*)[3][CEED_Q_VLA])in[0], (*q_data)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[1];
 
@@ -152,27 +163,18 @@ CEED_QFUNCTION(ElasFSInitialNH1F)(void *ctx, CeedInt Q, const CeedScalar *const 
   CeedScalar(*grad_u)[3][CEED_Q_VLA] = (CeedScalar(*)[3][CEED_Q_VLA])out[1];
 
   // Context
-  const Physics    context = (Physics)ctx;
-  const CeedScalar E       = context->E;
-  const CeedScalar nu      = context->nu;
-  const CeedScalar TwoMu   = E / (1 + nu);
-  const CeedScalar mu      = TwoMu / 2;
-  const CeedScalar Kbulk   = E / (3 * (1 - 2 * nu));  // Bulk Modulus
-  const CeedScalar lambda  = (3 * Kbulk - TwoMu) / 3;
+  const Physics_MR context = (Physics_MR)ctx;
+  const CeedScalar mu_1    = context->mu_1;
+  const CeedScalar mu_2    = context->mu_2;
+  const CeedScalar lambda  = context->lambda;
 
   // Formulation Terminology:
   //  I3    : 3x3 Identity matrix
   //  C     : right Cauchy-Green tensor
   //  C_inv : inverse of C
   //  F     : deformation gradient
-  //  S     : 2nd Piola-Kirchhoff (in current config)
-  //  P     : 1st Piola-Kirchhoff (in referential config)
-  // Formulation:
-  //  F =  I3 + grad_ue
-  //  J = det(F)
-  //  C = F(^T)*F
-  //  S = mu*I3 + (lambda*log(J)-mu)*C_inv;
-  //  P = F*S
+  //  S     : 2nd Piola-Kirchhoff
+  //  P     : 1st Piola-Kirchhoff
 
   // Quadrature Point Loop
   CeedPragmaSIMD for (CeedInt i = 0; i < Q; i++) {
@@ -208,14 +210,15 @@ CEED_QFUNCTION(ElasFSInitialNH1F)(void *ctx, CeedInt Q, const CeedScalar *const 
         {grad_u[2][0][i],     grad_u[2][1][i],     grad_u[2][2][i] + 1}
     };
 
-    // Common components of finite strain calculations
-    CeedScalar       Swork[6], Cinvwork[6], logJ;
     const CeedScalar tempgradu[3][3] = {
         {grad_u[0][0][i], grad_u[0][1][i], grad_u[0][2][i]},
         {grad_u[1][0][i], grad_u[1][1][i], grad_u[1][2][i]},
         {grad_u[2][0][i], grad_u[2][1][i], grad_u[2][2][i]}
     };
-    commonFS(lambda, mu, tempgradu, Swork, Cinvwork, &logJ);
+
+    // Common components of finite strain calculations
+    CeedScalar Swork[6], Cwork[6], Cinvwork[6], logJ;
+    commonFSMR(mu_1, mu_2, lambda, tempgradu, Swork, Cwork, Cinvwork, &logJ);
 
     // Second Piola-Kirchhoff (S)
     const CeedScalar S[3][3] = {
@@ -242,13 +245,13 @@ CEED_QFUNCTION(ElasFSInitialNH1F)(void *ctx, CeedInt Q, const CeedScalar *const 
     }
   }  // End of Quadrature Point Loop
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 // -----------------------------------------------------------------------------
 // Jacobian evaluation for hyperelasticity, finite strain
 // -----------------------------------------------------------------------------
-CEED_QFUNCTION(ElasFSInitialNH1dF)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
+CEED_QFUNCTION(ElasFSJacobian_MR)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
   // Inputs
   const CeedScalar(*deltaug)[3][CEED_Q_VLA] = (const CeedScalar(*)[3][CEED_Q_VLA])in[0],
         (*q_data)[CEED_Q_VLA]               = (const CeedScalar(*)[CEED_Q_VLA])in[1];
@@ -259,15 +262,10 @@ CEED_QFUNCTION(ElasFSInitialNH1dF)(void *ctx, CeedInt Q, const CeedScalar *const
   CeedScalar(*deltadvdX)[3][CEED_Q_VLA] = (CeedScalar(*)[3][CEED_Q_VLA])out[0];
 
   // Context
-  const Physics    context = (Physics)ctx;
-  const CeedScalar E       = context->E;
-  const CeedScalar nu      = context->nu;
-
-  // Constants
-  const CeedScalar TwoMu  = E / (1 + nu);
-  const CeedScalar mu     = TwoMu / 2;
-  const CeedScalar Kbulk  = E / (3 * (1 - 2 * nu));  // Bulk Modulus
-  const CeedScalar lambda = (3 * Kbulk - TwoMu) / 3;
+  const Physics_MR context = (Physics_MR)ctx;
+  const CeedScalar mu_1    = context->mu_1;
+  const CeedScalar mu_2    = context->mu_2;
+  const CeedScalar lambda  = context->lambda;
 
   // Quadrature Point Loop
   CeedPragmaSIMD for (CeedInt i = 0; i < Q; i++) {
@@ -288,6 +286,7 @@ CEED_QFUNCTION(ElasFSInitialNH1dF)(void *ctx, CeedInt Q, const CeedScalar *const
     // Compute graddeltau
     //   dXdx = (dx/dX)^(-1)
     // Apply dXdx to deltadu = graddelta
+    // this is dF
     CeedScalar graddeltau[3][3];
     for (CeedInt j = 0; j < 3; j++) {    // Component
       for (CeedInt k = 0; k < 3; k++) {  // Derivative
@@ -304,77 +303,99 @@ CEED_QFUNCTION(ElasFSInitialNH1dF)(void *ctx, CeedInt Q, const CeedScalar *const
         {grad_u[2][0][i],     grad_u[2][1][i],     grad_u[2][2][i] + 1}
     };
 
-    // Common components of finite strain calculations
-    CeedScalar       Swork[6], Cinvwork[6], logJ;
     const CeedScalar tempgradu[3][3] = {
         {grad_u[0][0][i], grad_u[0][1][i], grad_u[0][2][i]},
         {grad_u[1][0][i], grad_u[1][1][i], grad_u[1][2][i]},
         {grad_u[2][0][i], grad_u[2][1][i], grad_u[2][2][i]}
     };
-    commonFS(lambda, mu, tempgradu, Swork, Cinvwork, &logJ);
 
-    // deltaE - Green-Lagrange strain tensor
+    // Common components of finite strain calculations
+    CeedScalar Swork[6], Cwork[6], Cinvwork[6], logJ;
+    commonFSMR(mu_1, mu_2, lambda, tempgradu, Swork, Cwork, Cinvwork, &logJ);
+
+    // dE - Green-Lagrange strain tensor
     const CeedInt indj[6] = {0, 1, 2, 1, 0, 0}, indk[6] = {0, 1, 2, 2, 2, 1};
-    CeedScalar    deltaEwork[6];
+    CeedScalar    dEwork[6];
     for (CeedInt m = 0; m < 6; m++) {
-      deltaEwork[m] = 0;
-      for (CeedInt n = 0; n < 3; n++) deltaEwork[m] += (graddeltau[n][indj[m]] * F[n][indk[m]] + F[n][indj[m]] * graddeltau[n][indk[m]]) / 2.;
+      dEwork[m] = 0;
+      for (CeedInt n = 0; n < 3; n++) dEwork[m] += (graddeltau[n][indj[m]] * F[n][indk[m]] + F[n][indj[m]] * graddeltau[n][indk[m]]) / 2.;
     }
-    CeedScalar deltaE[3][3] = {
-        {deltaEwork[0], deltaEwork[5], deltaEwork[4]},
-        {deltaEwork[5], deltaEwork[1], deltaEwork[3]},
-        {deltaEwork[4], deltaEwork[3], deltaEwork[2]}
+    CeedScalar dE[3][3] = {
+        {dEwork[0], dEwork[5], dEwork[4]},
+        {dEwork[5], dEwork[1], dEwork[3]},
+        {dEwork[4], dEwork[3], dEwork[2]}
     };
-
     // C : right Cauchy-Green tensor
     // C^(-1) : C-Inverse
+    const CeedScalar C[3][3] = {
+        {Cwork[0], Cwork[5], Cwork[4]},
+        {Cwork[5], Cwork[1], Cwork[3]},
+        {Cwork[4], Cwork[3], Cwork[2]}
+    };
     const CeedScalar C_inv[3][3] = {
         {Cinvwork[0], Cinvwork[5], Cinvwork[4]},
         {Cinvwork[5], Cinvwork[1], Cinvwork[3]},
         {Cinvwork[4], Cinvwork[3], Cinvwork[2]}
     };
+    // -- C_inv:dE
+    CeedScalar Cinv_contract_dE = 0;
+    for (CeedInt j = 0; j < 3; j++) {
+      for (CeedInt k = 0; k < 3; k++) Cinv_contract_dE += C_inv[j][k] * dE[j][k];
+    }
 
+    // -- C:dE
+    CeedScalar C_contract_dE = 0;
+    for (CeedInt j = 0; j < 3; j++) {
+      for (CeedInt k = 0; k < 3; k++) C_contract_dE += C[j][k] * dE[j][k];
+    }
+
+    // -- dE*C_inv
+    CeedScalar dE_Cinv[3][3];
+    for (CeedInt j = 0; j < 3; j++) {
+      for (CeedInt k = 0; k < 3; k++) {
+        dE_Cinv[j][k] = 0;
+        for (CeedInt m = 0; m < 3; m++) dE_Cinv[j][k] += dE[j][m] * C_inv[m][k];
+      }
+    }
+
+    // -- C_inv*dE*C_inv
+    CeedScalar Cinv_dE_Cinv[3][3];
+    // This product is symmetric and we only use the upper-triangular part below, but naively compute the whole thing here
+    for (CeedInt j = 0; j < 3; j++) {
+      for (CeedInt k = 0; k < 3; k++) {
+        Cinv_dE_Cinv[j][k] = 0;
+        for (CeedInt m = 0; m < 3; m++) Cinv_dE_Cinv[j][k] += C_inv[j][m] * dE_Cinv[m][k];
+      }
+    }
+
+    // Compute dS = (mu_2)*((2*I_3:dE)*I_3 - dE) + 2*d*Cinv_dE_Cinv + lambda*Cinv_contract_dE*Cinvwork - 2*lambda*logJ*Cinv_dE_Cinv;
+    // (2*I_3:dE)*I_3 - dE = 2*trace(dE)*I_3 - dE = 2trace(dE) - dE on the diagonal
+    // (2*I_3:dE)*I_3 - dE = -dE elsewhere
+    // CeedScalar J = Jm1 + 1;
+    CeedScalar tr_dE = dE[0][0] + dE[1][1] + dE[2][2];
+    CeedScalar dSwork[6];
+    for (CeedInt i = 0; i < 6; i++) {
+      dSwork[i] = lambda * Cinv_contract_dE * Cinvwork[i] + 2 * (mu_1 + 2 * mu_2 - lambda * logJ) * Cinv_dE_Cinv[indj[i]][indk[i]] +
+                  2 * mu_2 * (tr_dE * (i < 3) - dEwork[i]);
+    }
+
+    CeedScalar dS[3][3] = {
+        {dSwork[0], dSwork[5], dSwork[4]},
+        {dSwork[5], dSwork[1], dSwork[3]},
+        {dSwork[4], dSwork[3], dSwork[2]}
+    };
     // Second Piola-Kirchhoff (S)
     const CeedScalar S[3][3] = {
         {Swork[0], Swork[5], Swork[4]},
         {Swork[5], Swork[1], Swork[3]},
         {Swork[4], Swork[3], Swork[2]}
     };
-
-    // deltaS = dSdE:deltaE
-    //      = lambda(C_inv:deltaE)C_inv + 2(mu-lambda*log(J))C_inv*deltaE*C_inv
-    // -- C_inv:deltaE
-    CeedScalar Cinv_contract_E = 0;
-    for (CeedInt j = 0; j < 3; j++) {
-      for (CeedInt k = 0; k < 3; k++) Cinv_contract_E += C_inv[j][k] * deltaE[j][k];
-    }
-    // -- deltaE*C_inv
-    CeedScalar deltaECinv[3][3];
+    // dP = dPdF:dF = dF*S + F*dS
+    CeedScalar dP[3][3];
     for (CeedInt j = 0; j < 3; j++) {
       for (CeedInt k = 0; k < 3; k++) {
-        deltaECinv[j][k] = 0;
-        for (CeedInt m = 0; m < 3; m++) deltaECinv[j][k] += deltaE[j][m] * C_inv[m][k];
-      }
-    }
-    // -- intermediate deltaS = C_inv*deltaE*C_inv
-    CeedScalar deltaS[3][3];
-    for (CeedInt j = 0; j < 3; j++) {
-      for (CeedInt k = 0; k < 3; k++) {
-        deltaS[j][k] = 0;
-        for (CeedInt m = 0; m < 3; m++) deltaS[j][k] += C_inv[j][m] * deltaECinv[m][k];
-      }
-    }
-    // -- deltaS = lambda(C_inv:deltaE)C_inv - 2(lambda*log(J)-mu)*(intermediate)
-    for (CeedInt j = 0; j < 3; j++) {
-      for (CeedInt k = 0; k < 3; k++) deltaS[j][k] = lambda * Cinv_contract_E * C_inv[j][k] - 2. * (lambda * logJ - mu) * deltaS[j][k];
-    }
-
-    // deltaP = dPdF:deltaF = deltaF*S + F*deltaS
-    CeedScalar deltaP[3][3];
-    for (CeedInt j = 0; j < 3; j++) {
-      for (CeedInt k = 0; k < 3; k++) {
-        deltaP[j][k] = 0;
-        for (CeedInt m = 0; m < 3; m++) deltaP[j][k] += graddeltau[j][m] * S[m][k] + F[j][m] * deltaS[m][k];
+        dP[j][k] = 0;
+        for (CeedInt m = 0; m < 3; m++) dP[j][k] += graddeltau[j][m] * S[m][k] + F[j][m] * dS[m][k];
       }
     }
 
@@ -382,18 +403,18 @@ CEED_QFUNCTION(ElasFSInitialNH1dF)(void *ctx, CeedInt Q, const CeedScalar *const
     for (CeedInt j = 0; j < 3; j++) {    // Component
       for (CeedInt k = 0; k < 3; k++) {  // Derivative
         deltadvdX[k][j][i] = 0;
-        for (CeedInt m = 0; m < 3; m++) deltadvdX[k][j][i] += dXdx[k][m] * deltaP[j][m] * wdetJ;
+        for (CeedInt m = 0; m < 3; m++) deltadvdX[k][j][i] += dXdx[k][m] * dP[j][m] * wdetJ;
       }
     }
   }  // End of Quadrature Point Loop
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 // -----------------------------------------------------------------------------
 // Strain energy computation for hyperelasticity, finite strain
 // -----------------------------------------------------------------------------
-CEED_QFUNCTION(ElasFSInitialNH1Energy)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
+CEED_QFUNCTION(ElasFSEnergy_MR)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
   // Inputs
   const CeedScalar(*ug)[3][CEED_Q_VLA] = (const CeedScalar(*)[3][CEED_Q_VLA])in[0], (*q_data)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[1];
 
@@ -401,13 +422,10 @@ CEED_QFUNCTION(ElasFSInitialNH1Energy)(void *ctx, CeedInt Q, const CeedScalar *c
   CeedScalar(*energy) = (CeedScalar(*))out[0];
 
   // Context
-  const Physics    context = (Physics)ctx;
-  const CeedScalar E       = context->E;
-  const CeedScalar nu      = context->nu;
-  const CeedScalar TwoMu   = E / (1 + nu);
-  const CeedScalar mu      = TwoMu / 2;
-  const CeedScalar Kbulk   = E / (3 * (1 - 2 * nu));  // Bulk Modulus
-  const CeedScalar lambda  = (3 * Kbulk - TwoMu) / 3;
+  const Physics_MR context = (Physics_MR)ctx;
+  const CeedScalar mu_1    = context->mu_1;
+  const CeedScalar mu_2    = context->mu_2;
+  const CeedScalar lambda  = context->lambda;
 
   // Quadrature Point Loop
   CeedPragmaSIMD for (CeedInt i = 0; i < Q; i++) {
@@ -424,7 +442,6 @@ CEED_QFUNCTION(ElasFSInitialNH1Energy)(void *ctx, CeedInt Q, const CeedScalar *c
         {q_data[4][i], q_data[5][i], q_data[6][i]},
         {q_data[7][i], q_data[8][i], q_data[9][i]}
     };
-
     // Compute grad_u
     //   dXdx = (dx/dX)^(-1)
     // Apply dXdx to du = grad_u
@@ -437,7 +454,7 @@ CEED_QFUNCTION(ElasFSInitialNH1Energy)(void *ctx, CeedInt Q, const CeedScalar *c
     }
 
     // E - Green-Lagrange strain tensor
-    //     E = 1/2 (grad_u + grad_u^T + grad_u^T*grad_u)
+    // E = 1/2 (grad_u + grad_u^T + grad_u^T*grad_u)
     const CeedInt indj[6] = {0, 1, 2, 1, 0, 0}, indk[6] = {0, 1, 2, 2, 2, 1};
     CeedScalar    E2work[6];
     for (CeedInt m = 0; m < 6; m++) {
@@ -449,21 +466,46 @@ CEED_QFUNCTION(ElasFSInitialNH1Energy)(void *ctx, CeedInt Q, const CeedScalar *c
         {E2work[5], E2work[1], E2work[3]},
         {E2work[4], E2work[3], E2work[2]}
     };
-    const CeedScalar Jm1  = computeJM1(grad_u);
-    const CeedScalar logJ = log1p_series_shifted(Jm1);
 
-    // Strain energy Phi(E) for compressible Neo-Hookean
-    energy[i] = (lambda * logJ * logJ / 2. - mu * logJ + mu * (E2[0][0] + E2[1][1] + E2[2][2]) / 2.) * wdetJ;
+    // C : right Cauchy-Green tensor
+    // C = I + 2E
+    const CeedScalar C[3][3] = {
+        {1 + E2[0][0], E2[0][1],     E2[0][2]    },
+        {E2[0][1],     1 + E2[1][1], E2[1][2]    },
+        {E2[0][2],     E2[1][2],     1 + E2[2][2]}
+    };
+    // Compute CC = C*C = C^2
+    CeedScalar CC[3][3];
+    for (CeedInt j = 0; j < 3; j++) {
+      for (CeedInt k = 0; k < 3; k++) {
+        CC[j][k] = 0;
+        for (CeedInt m = 0; m < 3; m++) CC[j][k] += C[j][m] * C[m][k];
+      }
+    }
+
+    const CeedScalar Jm1 = computeJM1(grad_u);
+    // CeedScalar J = Jm1 + 1;
+    // Compute invariants
+    // I_1 = trace(C)
+    const CeedScalar I_1 = C[0][0] + C[1][1] + C[2][2];
+    // Trace(C^2)
+    const CeedScalar tr_CC = CC[0][0] + CC[1][1] + CC[2][2];
+    // I_2 = 0.5(I_1^2 - trace(C^2))
+    const CeedScalar I_2 = 0.5 * (I_1 * I_1 - tr_CC);
+
+    const CeedScalar logJ = log1p_series_shifted(Jm1);
+    // Strain energy Phi(E) for Mooney-Rivlin
+    energy[i] = (0.5 * lambda * (logJ) * (logJ) - (mu_1 + 2 * mu_2) * logJ + (mu_1 / 2.) * (I_1 - 3) + (mu_2 / 2.) * (I_2 - 3)) * wdetJ;
 
   }  // End of Quadrature Point Loop
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 
 // -----------------------------------------------------------------------------
 // Nodal diagnostic quantities for hyperelasticity, finite strain
 // -----------------------------------------------------------------------------
-CEED_QFUNCTION(ElasFSInitialNH1Diagnostic)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
+CEED_QFUNCTION(ElasFSDiagnostic_MR)(void *ctx, CeedInt Q, const CeedScalar *const *in, CeedScalar *const *out) {
   // Inputs
   const CeedScalar(*u)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[0], (*ug)[3][CEED_Q_VLA] = (const CeedScalar(*)[3][CEED_Q_VLA])in[1],
         (*q_data)[CEED_Q_VLA] = (const CeedScalar(*)[CEED_Q_VLA])in[2];
@@ -472,13 +514,10 @@ CEED_QFUNCTION(ElasFSInitialNH1Diagnostic)(void *ctx, CeedInt Q, const CeedScala
   CeedScalar(*diagnostic)[CEED_Q_VLA] = (CeedScalar(*)[CEED_Q_VLA])out[0];
 
   // Context
-  const Physics    context = (Physics)ctx;
-  const CeedScalar E       = context->E;
-  const CeedScalar nu      = context->nu;
-  const CeedScalar TwoMu   = E / (1 + nu);
-  const CeedScalar mu      = TwoMu / 2;
-  const CeedScalar Kbulk   = E / (3 * (1 - 2 * nu));  // Bulk Modulus
-  const CeedScalar lambda  = (3 * Kbulk - TwoMu) / 3;
+  const Physics_MR context = (Physics_MR)ctx;
+  const CeedScalar mu_1    = context->mu_1;
+  const CeedScalar mu_2    = context->mu_2;
+  const CeedScalar lambda  = context->lambda;
 
   // Quadrature Point Loop
   CeedPragmaSIMD for (CeedInt i = 0; i < Q; i++) {
@@ -536,12 +575,37 @@ CEED_QFUNCTION(ElasFSInitialNH1Diagnostic)(void *ctx, CeedInt Q, const CeedScala
     for (CeedInt j = 0; j < 3; j++) {
       for (CeedInt m = 0; m < 3; m++) diagnostic[5][i] += E2[j][m] * E2[m][j] / 4.;
     }
-    diagnostic[6][i] = Jm1 + 1.;
+    diagnostic[6][i] = Jm1 + 1;
+
+    // C : right Cauchy-Green tensor
+    // C = I + 2E
+    const CeedScalar C[3][3] = {
+        {1 + E2[0][0], E2[0][1],     E2[0][2]    },
+        {E2[0][1],     1 + E2[1][1], E2[1][2]    },
+        {E2[0][2],     E2[1][2],     1 + E2[2][2]}
+    };
+    // Compute CC = C*C = C^2
+    CeedScalar CC[3][3];
+    for (CeedInt j = 0; j < 3; j++) {
+      for (CeedInt k = 0; k < 3; k++) {
+        CC[j][k] = 0;
+        for (CeedInt m = 0; m < 3; m++) CC[j][k] += C[j][m] * C[m][k];
+      }
+    }
+
+    // CeedScalar J = Jm1 + 1;
+    // Compute invariants
+    // I_1 = trace(C)
+    const CeedScalar I_1 = C[0][0] + C[1][1] + C[2][2];
+    // Trace(C^2)
+    const CeedScalar tr_CC = CC[0][0] + CC[1][1] + CC[2][2];
+    // I_2 = 0.5(I_1^2 - trace(C^2))
+    const CeedScalar I_2 = 0.5 * (pow(I_1, 2) - tr_CC);
 
     // Strain energy
-    diagnostic[7][i] = (lambda * logJ * logJ / 2. - mu * logJ + mu * (E2[0][0] + E2[1][1] + E2[2][2]) / 2.);
+    diagnostic[7][i] = (0.5 * lambda * logJ * logJ - (mu_1 + 2 * mu_2) * logJ + (mu_1 / 2.) * (I_1 - 3) + (mu_2 / 2.) * (I_2 - 3));
   }  // End of Quadrature Point Loop
 
-  return 0;
+  return CEED_ERROR_SUCCESS;
 }
 // -----------------------------------------------------------------------------
